# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Chronicle API
  version: 1.0.0
  description: REST API for Chronicle, a self-hosted TTRPG worldbuilding platform. Provides programmatic access to campaign data for external tool integration (Foundry VTT, custom scripts).
servers:
  - url: "{baseUrl}/api/v1"
    variables:
      baseUrl:
        default: http://localhost:8080
        description: Chronicle server base URL
security:
  - BearerApiKey: []
...
paths:
...
  /campaigns/{campaignId}/media:
    get:
      tags: [Media]
      summary: List media files
      operationId: listMedia
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated media list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MediaFile"
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Media]
      summary: Upload media file
      operationId: uploadMedia
      description: Max file size 10 MB.
      parameters:
        - $ref: "#/components/parameters/campaignId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                usage_type:
                  type: string
                  enum: [attachment, entity_image, avatar, backdrop]
                  default: attachment
      responses:
        "201":
          description: Uploaded media file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/media/stats:
    get:
      tags: [Media]
      summary: Get media storage stats
      operationId: getMediaStats
      parameters:
        - $ref: "#/components/parameters/campaignId"
      responses:
        "200":
          description: Storage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_files:
                    type: integer
                  total_bytes:
                    type: integer
                    format: int64
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /campaigns/{campaignId}/media/{mediaId}:
    get:
      tags: [Media]
      summary: Get media file details
      operationId: getMedia
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mediaId"
      responses:
        "200":
          description: Media file metadata with signed URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      tags: [Media]
      summary: Delete media file
      operationId: deleteMedia
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mediaId"
      responses:
        "204":
          description: Media deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # --- Maps ---
...
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: >
        API key passed as Bearer token. Keys are scoped to a single campaign
        with permission levels: read, write, sync. Rate limited per key.
  parameters:
    campaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
...
    mediaId:
      name: mediaId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
        type:
          type: string
        message:
          type: string
...
    MediaFile:
      type: object
      properties:
        id:
          type: string
        campaign_id:
          type: string
          nullable: true
        uploaded_by:
          type: string
        original_name:
          type: string
        mime_type:
          type: string
        file_size:
          type: integer
          format: int64
        usage_type:
          type: string
        url:
          type: string
          description: Signed URL for file access
        thumbnail_url:
          type: string
        thumbnails:
          type: object
          additionalProperties:
            type: string
          description: Map of size to signed thumbnail URL
        created_at:
          type: string
          format: date-time

    # --- Maps ---
...